Update README.md
Canadian Passport 


A Canadian passport serves as proof of the identity, nationality and status of the holder in Canada. According to the Passport Index, it ranks 9th in the world in terms of free movement and travel. The first Canadian passport took the form of a petition from the Governor-General of Canada.    
A Canadian passport allows the holder to leave and enter Canada, travel to other countries in accordance with visa requirements, facilitate the process of assistance from Canadian consular officials and to seek protection from the holder where necessary. A Canadian passport is issued by the Passport Program of Immigration, Refugees and Citizenship Canada (IRCC). Anyone with a valid Canadian passport can visit the United States (USA), including a visit.    
At the arrival of travelers at a Canadian port of entry, the CBSA Border Service Officer (BSO) must show that they have met the requirements for entry into Canada. If you do not have the right documents, you may be denied entry to Canada. Permanent residents of Canada and members of the Nexus Fast Program who travel without a passport or proof of permanent residence are requested to present them to the BSO upon arrival at the border.    
You must have a valid Canadian passport and a certified visa before boarding your flight. Upon entering Canada, a border official will inspect your passport and valid visa when you enter Canada, but no one is required to do so. What you need depends on where you are, who you are traveling with and what documents you are traveling with.    
The validity of the e-passport for adults is 5-10 years from the first renewal application. Secondly, a genuine passport opens the door to something that is closed to many citizens of a country.    
A genuine Canadian passport has an RFID chip and comprises 36 pages (29 pages for visa labels and stamps). A Canadian passport from Canada is a duplicate of your fake Canadian passport with no discernible differences. The US offers a guarantee for a fake Canadian passport for $400 and a genuine Canadian passport for $800 for every one.    
Talk to us and we will make sure that you receive a fake Canadian passport in good time so that you can implement your travel plans without any problems. Our specialists produce your Canadian passport with special equipment and state-of-the-art technology and procedures. The Canadian passport you receive from us contains all important data and information such as your name, date of birth, photo and signature.    
If you create a fake passport in Canada or the USA, it is essential that you use all essential security features. When you search online for a fake Canadian passport, you must provide all relevant details such as name, age, gender and address when you are done.    
It is best to apply to have your genuine Canadian passport restored. In particular, you should wait until you show a 21 + ID. Buy counterfeit Canadian ID here.    
Investigators investigating the Al Qaeda terror network have pointed to a common thread: the ease with which many members of the group travel with forged passports and illegally obtained immigration documents around the world. Ahmed Ressam, who is at the centre of a Frontline report, is an expert on fake passports.    
Ahmed Ressam, who is in the center of this FRONTLINE report, used a forged French passport to enter Canada and apply for political asylum. When caught at the Canadian-American border, he wore a fake Canadian passport under the alias Benni Noris. Investigators discovered he had entered Canada from France on a fake passport.    
Ahmed Ressam had manipulated the Canadian asylum system to claim political asylum at the Canadian-American border knowing that Canadian authorities would refer his case to an authority that would allow him to stay in Canada while he awaits his asylum trial. He later testified that an Al-Qaeda representative had recruited him to supply the group with false Canadian passports.